--- src/features/home/components/LandingPageNew/LandingPageNew.module.css
@@ -1,179 +1,0 @@
-.hero {
-  min-height: 100vh; /* Fallback */
-  min-height: 100dvh; /* Modern viewport units */
-  height: var(--tg-viewport-stable-height);
-  width: 100vw;
-  overflow-x: hidden;
-  padding: var(--spacing-md) var(--spacing-sm);
-  background-color: var(--app-background);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: relative;
-  transition: background-color var(--animation-duration-normal) var(--animation-easing);
-  animation: fadeIn var(--animation-duration-normal) var(--animation-easing);
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start;
-  text-align: center;
-  gap: var(--spacing-md);
-  width: 100%;
-  max-width: 100%;
-  padding: var(--spacing-md) var(--spacing-xs);
-}
-
-.logo {
-  max-width: min(267px, 60vw);
-  width: 100%;
-  height: auto;
-  display: block;
-  margin-bottom: var(--spacing-md);
-  object-fit: contain;
-  transition: all var(--animation-duration-normal) var(--animation-easing);
-  filter: drop-shadow(0 0 10px var(--shadow-primary));
-  z-index: 1;
-  position: relative;
-}
-
-.carousel {
-  width: 100%;
-  margin: var(--spacing-md) 0;
-}
-
-.buttonContainer {
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  align-items: center;
-  font-size: 16px;
-  color: #000;
-  text-align: center;
-  line-height: 1;
-  justify-content: center;
-  padding: 16px 0;
-}
-
-.buttonIcon {
-  width: 16px;
-  height: 16px;
-  display: block;
-}
-
-.accountSection {
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  align-items: center;
-  justify-content: start;
-  padding: 16px 0;
-}
-
-.accountText {
-  color: var(--color-text-light);
-  font: 400 12px/2 Inter, sans-serif;
-}
-
-[data-theme="dark"] .accountText {
-  color: var(--color-text-dark);
-}
-
-.accountLink {
-  display: flex;
-  margin-top: 8px;
-  align-items: center;
-  gap: 8px;
-  font-size: 16px;
-  color: var(--color-primary-dark);
-  justify-content: start;
-  text-decoration: none;
-  transition: color var(--animation-duration-normal) var(--animation-easing);
-}
-
-[data-theme="dark"] .accountLink {
-  color: var(--color-primary-light);
-}
-
-.accountLink:hover {
-  color: var(--color-primary);
-}
-
-.linkLabel {
-  font-family: Inter, sans-serif;
-  font-weight: 600;
-  line-height: 1;
-  align-self: stretch;
-  margin: auto 0;
-}
-
-.linkIcon {
-  width: 15px;
-  height: 24px;
-  margin: auto 0;
-  display: block;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-/* iPhone SE and similar small screens */
-@media (max-width: 375px) {
-  .hero {
-    padding: var(--spacing-sm) var(--spacing-xs);
-  }
-
-  .container {
-    gap: var(--spacing-sm);
-    padding: var(--spacing-sm) var(--spacing-xs);
-  }
-
-  .logo {
-    max-width: min(200px, 50vw);
-    margin-bottom: var(--spacing-sm);
-  }
-
-  .buttonContainer {
-    padding: 12px 0;
-  }
-
-  .accountSection {
-    padding: 12px 0;
-  }
-}
-
-/* Tablet and up */
-@media (min-width: 576px) {
-  .hero {
-    padding: var(--spacing-xl) var(--spacing-md);
-  }
-
-  .container {
-    gap: var(--spacing-xl);
-    padding: var(--spacing-xl) var(--spacing-md);
-    max-width: 768px;
-  }
-
-  .logo {
-    margin-bottom: var(--spacing-xl);
-  }
-
-  .buttonContainer {
-    max-width: 500px;
-  }
-}
-
-/* Desktop and up */
-@media (min-width: 992px) {
-  .container {
-    max-width: 1200px;
-  }
-}
