--- src/features/home/components/LandingPageNew/index.js
@@ -1,1 +1,0 @@
-export { default as LandingPageNew } from './LandingPageNew';
