+++ src/features/home/index.js
@@ -1,3 +1,2 @@
 export * from './components/LandingPage';
 export * from './components/Dashboard';
-export * from './components/LandingPageNew';
